[CSL-2149] checkCoin checkCoinPortion remove tests
Those tests are not needed. They only checked that the "smart
constructor" didn't do anything ridiculous i.e. change the value. Now we
don't have smart constructors, we have verifiers. Obviously the actual
constructor (newtype) doesn't change the value, so all we have to check
is that the verifier correctly rejects invalid values. A test for that
seems a bit silly: for every number over the limit, test that the
validator rejects it... well, ok, I can prove that very easily, so why
bother?